+++ CHANGES.rst
@@ -11,6 +11,7 @@ Version history
 - Fixed ``AttributeError`` when metadata contains user defined column types
 - Fixed ``AssertionError`` when metadata contains a column type that is a type decorator
   with an all-uppercase name
+- Fixed MySQL ``DOUBLE`` column types being rendered with the wrong arguments
 
 **3.0.0rc5**